refactor issue components
