+++ frontend/src/components/Issue/styles.js
