--- frontend/src/components/IssueList/index.js
@@ -1,3 +1,0 @@
-import FilteredIssueList from './FilteredIssueList';
-
-export {FilteredIssueList};
